Change .env.example

+++ resources/views/journals/journals/action.blade.php
@@ -6,8 +6,10 @@
 @can('journals.show')
   <a href="{{ route('journals.show', $uuid) }}" class="text-primary me-2"><i class="fa fa-sm fa-eye"></i></a>
 @endcan
-@if (!$is_approved)
-  @can('journals.destroy')
-    <a href="#" onclick="deleteJournal(`{{ route('journals.destroy', $uuid) }}`)" class="text-danger me-2"><i class="fa fa-sm fa-trash"></i></a>
-  @endcan
+@if(isRoleName() === Constant::ADMIN)
+  @if (!$is_approved)
+    @can('journals.destroy')
+      <a href="#" onclick="deleteJournal(`{{ route('journals.destroy', $uuid) }}`)" class="text-danger me-2"><i class="fa fa-sm fa-trash"></i></a>
+    @endcan
+  @endif
 @endif
